server/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
groq
requests
gunicorn